[loaders] Split by whether or not authentication is required.
