+++ lib/loaders/per_type/followed_artist.js
