+++ lib/loaders/per_type/followed_profile.js
@@ -1,6 +1,6 @@
 import DataLoader from "dataloader"
-import gravity from "../apis/gravity"
-import { toKey } from "../helpers"
+import gravity from "lib/apis/gravity"
+import { toKey } from "lib/helpers"
 import { map, find, extend, first } from "lodash"
 
 const followedProfileLoader = new DataLoader(
