+++ lib/loaders/per_type/saved_artwork.js
