+++ schema/artist/index.js
@@ -4,7 +4,7 @@ import { pageable, getPagingParameters } from "relay-cursor-paging"
 import { connectionFromArraySlice, connectionDefinitions } from "graphql-relay"
 import { assign, compact, defaults, first, has } from "lodash"
 import { exclude } from "lib/helpers"
-import followedArtistLoader from "lib/loaders/followed_artist"
+import followedArtistLoader from "lib/loaders/per_type/followed_artist"
 import cached from "schema/fields/cached"
 import initials from "schema/fields/initials"
 import { markdown, formatMarkdownValue } from "schema/fields/markdown"
